Refactor memes and timekeeping commands
 - Move !belopa and !heresy to the memes section.
 - Move !order to it's own module, memes.order.
 - Add documentation for !march and !truemarch.
 - Update the start time of Desert Bus to 2021.
 - Update the !plan for Julie's stream.
